Refactor API Controllers and Improve Logging
- Updated logging statements in PasswordController, PhonebookController, PublipostageController, UserController to improve readability.
- Enhanced error handling and validation in PhonebookController and PublipostageController.
- Modified the ContactImport class to handle imports more effectively and added support for publishing status.
- Updated API routes for better organization and clarity, including new routes for importing contacts in Phonebook and Publipostage.
- Improved API documentation to reflect changes in endpoints and request/response structures.
- Removed unused 'status' column from contacts table migration.
- Added new success messages for contact import in English and Portuguese language files.

+++ app/Http/Controllers/API/PhonebookController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers\API;
 
 use Excel;
 use \Carbon\Carbon;
-use Illuminate\Support\Str;
 use App\Imports\ContactImport;
 use Illuminate\Validation\Rule;
 use Illuminate\Http\{Request, JsonResponse};
@@ -28,7 +27,7 @@ class PhonebookController extends BaseController
     * )
     */
     public function index(Request $request): JsonResponse {
-        //User
+        // User
         $user = Auth::user();
 		App::setLocale($user->lg);
         try {
@@ -38,13 +37,11 @@ class PhonebookController extends BaseController
             // Code to list contacts
             $query = Contact::select('uid', 'label', 'number', 'gender', 'date_at', 'field1', 'field2', 'field3');
             if ($search) $query->where('label', 'LIKE', '%'.$search.'%');
-            $query->where('status', 0)
+            $query->where('user_id', $user->id)
             ->where('blacklist', 0)
             ->where('publipostage', 0)
-            ->where('user_id', $user->id)
             ->orderByDesc('created_at')
             ->get();
-            $total = $query->count();
             $contacts = $query->paginate($limit, ['*'], 'page', $num);
             // Vérifier si les données existent
             if ($contacts->isEmpty()) {
@@ -56,13 +53,18 @@ class PhonebookController extends BaseController
                 'uid' => $data->uid,
                 'label' => $data->label,
                 'number' => $data->number,
-                'gender' => $data->gender,
-                'date_at' => Carbon::parse($data->date_at)->format('d/m/Y'),
-                'field1' => $data->field1,
-                'field2' => $data->field2,
-                'field3' => $data->field3,
+                'gender' => $data->gender ?? '',
+                'date_at' => $data->date_at != null ? Carbon::parse($data->date_at)->format('d/m/Y'):'',
+                'field1' => $data->field1 ?? '',
+                'field2' => $data->field2 ?? '',
+                'field3' => $data->field3 ?? '',
             ]);
-            return $this->sendSuccess(__('message.listcontact'), $data);
+            $total = [
+                'currentPage' => $contacts->currentPage(),
+                'lastPage' => $contacts->lastPage(),
+                'total' => $contacts->total(),
+            ];
+            return $this->sendSuccess(__('message.listcontact'), $data, 200, $total);
         } catch (\Exception $e) {
             Log::warning("Contact::index - Erreur d'affichage de Contacts: " . $e->getMessage());
             return $this->sendError(__('message.error'));
@@ -95,10 +97,10 @@ class PhonebookController extends BaseController
     * )
     */
     public function store(Request $request): JsonResponse {
-        //User
+        // User
         $user = Auth::user();
 		App::setLocale($user->lg);
-        //Validator
+        // Validator
         $validator = Validator::make($request->all(), [
             'label' => 'required',
             'number' => [
@@ -106,7 +108,7 @@ class PhonebookController extends BaseController
                 'digits:9',
                 'numeric',
                 Rule::unique('contacts')->where(function ($query) use ($user) {
-                    return $query->where('user_id', $user->id)->where('publipostage', 0)->where('status', 0);
+                    return $query->where('user_id', $user->id)->where('publipostage', 0);
                 }),
             ],
             'gender' => 'present',
@@ -115,15 +117,15 @@ class PhonebookController extends BaseController
             'field2' => 'present',
             'field3' => 'present',
         ]);
-        //Error field
+        // Error field
         if ($validator->fails()) {
-            Log::warning("Contact::store - Validator : " . $validator->errors()->first() . " - ".json_encode($request->all()));
+            Log::warning("Contact::store - Validator : " . $validator->errors()->first() . " - " . json_encode($request->all()));
             return $this->sendError(__('message.fielderr'), $validator->errors()->first(), 422);
         }
         // Vérifier du préfixe téléphonique
         $prefix = Prefix::where('label', substr($request->number, 0, 2))->first();
         if (!$prefix) {
-            Log::warning("Contact::store - Validator number : ".json_encode($request->all()));
+            Log::warning("Contact::store - Validator number : " . json_encode($request->all()));
             return $this->sendError(__('message.numbernot'), [], 422);
         }
         // Création de la reclamation
@@ -176,10 +178,10 @@ class PhonebookController extends BaseController
     * )
     */
     public function update(request $request, $uid): JsonResponse {
-        //User
+        // User
         $user = Auth::user();
 		App::setLocale($user->lg);
-        //Validator
+        // Validator
         $validator = Validator::make($request->all(), [
             'label' => 'required',
             'number' => [
@@ -187,7 +189,7 @@ class PhonebookController extends BaseController
                 'digits:9',
                 'numeric',
                 Rule::unique('contacts')->where(function ($query) use ($user) {
-                    return $query->where('user_id', $user->id)->where('publipostage', 0)->where('status', 0);
+                    return $query->where('user_id', $user->id)->where('publipostage', 0);
                 }),
             ],
             'gender' => 'present',
@@ -196,15 +198,15 @@ class PhonebookController extends BaseController
             'field2' => 'present',
             'field3' => 'present',
         ]);
-        //Error field
+        // Error field
         if ($validator->fails()) {
-            Log::warning("Contact::update - Validator : " . $validator->errors()->first() . " - ".json_encode($request->all()));
+            Log::warning("Contact::update - Validator : " . $validator->errors()->first() . " - " . json_encode($request->all()));
             return $this->sendError(__('message.fielderr'), $validator->errors(), 422);
         }
         // Vérifier du préfixe téléphonique
         $prefix = Prefix::where('label', substr($request->number, 0, 2))->first();
         if (!$prefix) {
-            Log::warning("Contact::update - Validator number : ".json_encode($request->all()));
+            Log::warning("Contact::update - Validator number : " . json_encode($request->all()));
             return $this->sendError(__('message.numbernot'), [], 422);
         }
         // Vérifier si l'ID est présent et valide
@@ -238,33 +240,54 @@ class PhonebookController extends BaseController
     // Suppression d'un Contact
     /**
     *   @OA\Delete(
-    *   path="/api/phonebooks/{uid}",
+    *   path="/api/phonebooks/delete",
     *   tags={"Phonebooks"},
     *   operationId="deleteContact",
-    *   description="Suppression d'un Contact",
+    *   description="Suppression d'un ou de plusieurs Contacts",
     *   security={{"bearer":{}}},
-    *   @OA\Response(response=201, description="Contact supprimé avec succès."),
+    *   @OA\RequestBody(
+    *      required=true,
+    *      @OA\JsonContent(
+    *         required={"contacts"},
+    *         @OA\Property(property="contacts", type="array", @OA\Items(
+    *               example="['910102034', '920102034', '930102034']"
+    *           )
+    *         ),
+    *      )
+    *   ),
+    *   @OA\Response(response=201, description="Contacts supprimés avec succès."),
     *   @OA\Response(response=400, description="Serveur indisponible."),
     *   @OA\Response(response=404, description="Page introuvable.")
     * )
     */
-    public function destroy($uid): JsonResponse {
-        //User
+    public function destroy(request $request): JsonResponse {
+        // User
         $user = Auth::user();
 		App::setLocale($user->lg);
+        // Validator
+        $validator = Validator::make($request->all(), [
+            'contacts' => 'required|array',
+        ]);
+        // Error field
+        if($validator->fails()){
+            Log::warning("Contact::destroy - Validator : " . $validator->errors()->first() . " - " . json_encode($request->all()));
+            return $this->sendError(__('message.fielderr'), $validator->errors(), 422);
+        }
         try {
-            // Vérification si le Contact est attribué à une demande
-            $contact = Contact::where('uid', $uid)->first();
-            // Suppression
-            $deleted = Contact::destroy($contact->id);
-            if (!$deleted) {
-                Log::warning("Contact::destroy - Tentative de suppression d'un Contact inexistante : " . $uid);
-                return $this->sendError(__('message.error'), [], 403);
-            }
-            $find = GroupContact::where('contact_id', $contact->id)->first();
-            if ($find) {
-                GroupContact::where('contact_id', $contact->id)->delete();
-            }
+            foreach ($request->contacts as $number) :
+                // Vérification si le Contact est attribué à une demande
+                $contact = Contact::where('number', $number)
+                ->where('user_id', $user->id)
+                ->where('publipostage', 0)
+                ->first();
+                // Suppression
+                $deleted = Contact::destroy($contact->id);
+                if ($deleted) {
+                    $find = GroupContact::where('contact_id', $contact->id)->first();
+                    $profil = Profile::find($user->profile_id);
+                    if ($find) $find->delete();
+                }
+            endforeach;
             return $this->sendSuccess(__('message.delcontact'), [], 201);
         } catch(\Exception $e) {
             Log::warning("Contact::destroy - Erreur lors de la suppression d'un Contact : " . $e->getMessage());
@@ -276,7 +299,7 @@ class PhonebookController extends BaseController
     * @OA\Post(
     *   path="/api/phonebooks/imports",
     *   tags={"Phonebooks"},
-    *   operationId="importContact",
+    *   operationId="importNum",
     *   description="Importation d'un Contact",
     *   security={{"bearer":{}}},
     *   @OA\RequestBody(
@@ -295,32 +318,30 @@ class PhonebookController extends BaseController
     * )
     */
     public function imports(Request $request): JsonResponse {
-        //User
+        // User
         $user = Auth::user();
 		App::setLocale($user->lg);
-        //Validator
+        // Validator
         $validator = Validator::make($request->all(), [
             'files' => 'required|file|mimes:xlsx,xls|max:2048',
         ]);
-        //Error field
+        // Error field
         if ($validator->fails()) {
-            Log::warning("Contact::store - Validator : " . $validator->errors()->first() . " - ".json_encode($request->all()));
+            Log::warning("Contact::imports - Validator : " . $validator->errors()->first() . " - " . json_encode($request->all()));
             return $this->sendError(__('message.fielderr'), $validator->errors()->first(), 422);
         }
-        $import = new ContactImport($user);
-    
+        $import = new ContactImport($user, 0);
         try {
             Excel::import($import, $request->file('files'));
-            
             $errors = $import->getErrors();
-            
             if (!empty($errors)) {
-                return $this->sendError("Certaines lignes n'ont pas pu être importées", $errors, 422);
+                Log::warning("Contact::imports : Certaines lignes n'ont pas pu être importées" . json_encode($errors));
+                return $this->sendError(__('message.fielderr'), [], 422);
             }
-            
-            return $this->sendSuccess('Import réussi', [], 201);
+            return $this->sendSuccess(__('message.impcontact'), [], 201);
         } catch (\Exception $e) {
-            return $this->sendError("Erreur lors de l'import", $e->getMessage(), 400);
+            Log::warning("Contact::imports : " . $e->getMessage());
+            return $this->sendError(__('message.fielderr'), [], 400);
         }
     }
 }
